refactor data collection schema
